+++ Modix.Data/Models/Moderation/InfractionSummary.cs
@@ -121,7 +121,7 @@ namespace Modix.Data.Models.Moderation
             = entity => new InfractionSummary()
             {
                 Id = entity.Id,
-                GuildId = (ulong)entity.GuildId,
+                GuildId = entity.GuildId,
                 // https://github.com/aspnet/EntityFrameworkCore/issues/12834
                 //Type = entity.Type,
                 Type = Enum.Parse<InfractionType>(entity.Type.ToString()),
